Add files via upload
This is my final project from my CS210 (Computer Science 3) course. This project takes in a list of items from a file and allows the user to sift through a menu to display specific data from this file. When the program closes, the program outputs a file of all the items in the list with a number next to it determining its frequency.